- refactored in to a "Logger" flavor for just pure collection and a "Viewer" flavor for real-time monitoring
