+++ torgi/build.gradle
@@ -26,6 +26,20 @@ android {
 //            path "CMakeLists.txt"
 //        }
 //    }
+    flavorDimensions "version"
+    productFlavors {
+        logger {
+            dimension "version"
+            applicationId "org.sofwerx.torgi.logger"
+            versionNameSuffix "-logger"
+        }
+        viewer {
+            dimension "version"
+            applicationId "org.sofwerx.torgi.viewer"
+            versionNameSuffix "-viewer"
+        }
+    }
+
     compileOptions {
         targetCompatibility 1.8
         sourceCompatibility 1.8
@@ -36,11 +50,11 @@ dependencies {
     //implementation fileTree(include: ['*.jar'], dir: 'libs')
     //implementation project(':geopackage-sdk')
 
-    implementation 'com.github.PhilJay:MPAndroidChart:v3.0.3' //for charts
-    implementation 'org.osmdroid:osmdroid-android:6.0.2'
+    viewerImplementation 'com.github.PhilJay:MPAndroidChart:v3.0.3' //for charts
+    viewerImplementation 'org.osmdroid:osmdroid-android:6.0.2'
     implementation 'mil.nga.geopackage:geopackage-android:3.0.2'
     implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+    viewerImplementation 'com.android.support.constraint:constraint-layout:1.1.3'
     //implementation 'com.google.android.gms:play-services-maps:15.0.1'
     implementation 'com.android.support:design:27.1.1'
     testImplementation 'junit:junit:4.12'
