+++ torgi/src/main/AndroidManifest.xml
@@ -14,10 +14,6 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
 
-        <meta-data
-            android:name="com.google.android.geo.API_KEY"
-            android:value="@string/google_maps_key" />
-
         <service android:name="org.sofwerx.torgi.service.TorgiService"/>
 
         <activity
@@ -27,15 +23,6 @@
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-            <!--android:theme="@style/AppTheme.NoActionBar">-->
-        </activity>
-        <activity
-            android:name=".ui.MonitorActivity"
-            android:label="TORGI Monitor">
-        </activity>
-        <activity
-            android:name=".ui.AboutActivity"
-            android:label="@string/about">
         </activity>
     </application>
 </manifest>
