--- torgi/src/main/res/layout/about_activity.xml
@@ -1,103 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".ui.AboutActivity"
-    android:background="@color/colorPrimaryDark"
-    android:padding="8dp">
-
-    <ImageView
-        android:id="@+id/aboutLogo"
-        android:layout_width="match_parent"
-        android:layout_height="50dp"
-        android:src="@mipmap/ic_launcher"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <ScrollView
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:fadingEdgeLength="40dp"
-        android:requiresFadingEdge="vertical"
-        app:layout_constraintTop_toBottomOf="@id/aboutLogo"
-        app:layout_constraintBottom_toTopOf="@id/aboutCopyright"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        tools:layout_editor_absoluteX="8dp">
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:animateLayoutChanges="true"
-            android:orientation="vertical">
-
-            <View
-                android:layout_width="match_parent"
-                android:layout_height="4dp"
-                android:layout_margin="8dp"
-                android:background="@android:color/white" />
-
-            <TextView
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:gravity="center"
-                android:text="@string/legal"
-                android:textColor="@android:color/white"
-                android:textStyle="bold" />
-
-            <TextView
-                android:id="@+id/legalAckTitle"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:drawableStart="@drawable/icon_expanded_white"
-                android:text="@string/acknowledgements"
-                android:textColor="@android:color/white" />
-
-            <TextView
-                android:id="@+id/legalAck"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:paddingStart="24dp"
-                android:text="@string/acknowledgements"
-                android:textColor="@android:color/white"
-                android:textColorLink="@android:color/white"
-                android:visibility="gone" />
-
-            <TextView
-                android:id="@+id/legalLicenseTitle"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:drawableStart="@drawable/icon_expanded_white"
-                android:text="@string/licenses"
-                android:textColor="@android:color/white" />
-
-            <TextView
-                android:id="@+id/legalLicense"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:paddingStart="24dp"
-                android:text="@string/acknowledgements"
-                android:textColor="@android:color/white"
-                android:textColorLink="@android:color/white"
-                android:visibility="gone" />
-
-        </LinearLayout>
-    </ScrollView>
-
-    <TextView
-        android:id="@+id/aboutCopyright"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:gravity="center"
-        android:text="@string/copyright"
-        android:textColor="@android:color/white"
-        android:textSize="10sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-
-</android.support.constraint.ConstraintLayout>
