--- torgi/src/main/res/layout/activity_monitor.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:theme="@style/AppTheme"
-    tools:context=".ui.MainActivity">
-
-    <TextView
-        android:id="@+id/monitorTextOverview"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:drawableStart="@drawable/icon_gnns"
-        android:gravity="center_vertical"
-        android:text="No fix" />
-
-    <com.github.mikephil.charting.charts.CombinedChart
-        android:id="@+id/chartEW"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="0.5"/>
-
-    <org.osmdroid.views.MapView
-        android:id="@+id/maposm"
-        android:layout_height="0dp"
-        android:layout_width="match_parent"
-        android:layout_weight="0.5"
-        tilesource="openstreetmaps" />
-</LinearLayout>
